# Conflicts:
#	.idea/artifacts/minvakt_war_exploded.xml
#	.idea/compiler.xml
#	src/main/webapp/js/employee.js

+++ src/main/webapp/js/employee.js
@@ -7,8 +7,6 @@
 var rootURL = "http://localhost:8080/api/employees/";
 
 $(document).ready(function() {
-  //FIXME: make this work without commenting out. Prob just move the functions on the bottom to another js file.
-/* 
     // When the user clicks on the create button, open the modal
     $("#createEmployee").on('click', function() {
         $("#createModal").modal('show');
@@ -16,21 +14,31 @@ $(document).ready(function() {
 
     // When the user clicks on the update button, open the modal
     $("#updateEmployee").on('click', function() {
-        $("#updateModal").modal('show');
-        $(".username_filler").html('<b>Edit user: ' + $('tr.selected td:eq(5)').text() + '</b>');
-        $("#inputFirstname").attr('value', $('tr.selected td:eq(1)').text());
-        $("#inputLastname").attr('value', $('tr.selected td:eq(0)').text());
-        $("#inputPhone").attr('value', $('tr.selected td:eq(2)').text());
-        $("#inputEmail").attr('value', $('tr.selected td:eq(3)').text());
-        $("#inputCategory").val($('tr.selected td:eq(4)').text());
+        if($('tr.selected td:eq(5)').text() != "") {
+            $("#updateModal").modal('show');
+            $(".username_filler").html('<b>Edit user: ' + $('tr.selected td:eq(5)').text() + '</b>');
+            $("#inputFirstname").attr('value', $('tr.selected td:eq(1)').text());
+            $("#inputLastname").attr('value', $('tr.selected td:eq(0)').text());
+            $("#inputPhone").attr('value', $('tr.selected td:eq(2)').text());
+            $("#inputEmail").attr('value', $('tr.selected td:eq(3)').text());
+            $("#inputCategory").val($('tr.selected td:eq(4)').text());
+        } else {
+            alert("Trykk på en ansatt før du endrer.");
+        }
+
     });
 
     $("#updateRoleAndPassword").on('click', function() {
-        $.getJSON('api/users/' + $('tr.selected td:eq(5)').text(), function(result) {
-            $("#updateRoleAndPasswordModal").modal('show');
-            $(".username_filler").html('<b>Edit user: ' + $('tr.selected td:eq(5)').text() + '</b>');
-            $("#inputRole1").val(result.role + "Sel");
-        });
+        if($('tr.selected td:eq(5)').text() != "") {
+            $.getJSON('api/users/' + $('tr.selected td:eq(5)').text(), function(result) {
+                $("#updateRoleAndPasswordModal").modal('show');
+                $(".username_filler").html('<b>Edit user: ' + $('tr.selected td:eq(5)').text() + '</b>');
+                $("#inputRole1").val(result.role + "Sel");
+            });
+        } else {
+            alert("Trykk på en ansatt før du endrer.");
+        }
+
     });
 
     // When the user clicks on <span> (x), close the modal
@@ -60,7 +68,7 @@ $(document).ready(function() {
             $("#updateRoleAndPasswordModal").modal('hide');
         }
     }
-*/
+
     function editEmployee() {
         var user_id = $('tr.selected td:eq(5)').text();
         console.log('editEmployee with user_id: ' + user_id);
@@ -86,6 +94,7 @@ $(document).ready(function() {
                 console.log("Error: " + textStatus);
             }
         });
+
     }
 
     //Hvis en bruker med samme brukernavn allerede finnes vil bare kontoen oppdateres med nytt passord og rolle(bruker/admin/inactive)
@@ -111,6 +120,7 @@ $(document).ready(function() {
             $("#inputPassword").focus();
             return false;
         }
+
         $.ajax({
             type: 'POST',
             contentType: 'application/json',
@@ -126,9 +136,9 @@ $(document).ready(function() {
 
                 //creating empty employee with same user_id
                 $.ajax({
-                    type: 'POST',
+                    type: 'PUT',
                     contentType: 'application/json',
-                    url: "api/employees",
+                    url: 'api/users/' + $("#inputUsername").val(),
                     dataType: "json",
                     data: JSON.stringify({
                         "user_id": $("#inputUsername").val(),
@@ -212,19 +222,23 @@ $(document).ready(function() {
 
     //delete employee
     $("#deleteEmployee").click(function () {
-        var user_id = $('tr.selected td:eq(5)').text();
-        console.log(user_id);
-        $.ajax({
-            url: rootURL + user_id,
-            type: 'DELETE',
-            success: function(result) {
-                console.log("Employee was deleted with user_id: " + user_id);
-                window.location.reload();
-            },
-            error: function(data, textStatus, jqXHR){
-                console.log("Error: " + textStatus);
-            }
-        });
+        if($('tr.selected td:eq(5)').text() != "") {
+            var user_id = $('tr.selected td:eq(5)').text();
+            $.ajax({
+                url: rootURL + user_id,
+                type: 'DELETE',
+                success: function(result) {
+                    console.log("Employee was deleted with user_id: " + user_id);
+                    window.location.reload();
+                },
+                error: function(data, textStatus, jqXHR){
+                    console.log("Error: " + textStatus);
+                }
+            });
+        } else {
+            alert("Trykk på en ansatt før du fjerner.");
+        }
+
     });
   
     $("#availableEmployeesTable").DataTable({
@@ -271,8 +285,8 @@ $(document).ready(function() {
                 console.log("Error: " + textStatus);
             }
         });
-
     }
+
     $("#availableEmployees").click(function () {
         var category = $("#category").val();
         var my_date = $("#my_date").val();
@@ -290,7 +304,7 @@ $(document).ready(function() {
         $.ajax({
             type: 'PUT',
             contentType: 'application/json',
-            url: "api/shift_lists/deviance/" + user_id+"/"+ shift_id,
+            url: "api/shift_lists/deviance",
             data: '{"user_id": "' + user_id + '", "shift_id" : "' + shift_id + '", "my_date" : "' + my_date + '", "deviance" : "' + deviance + '"}',
             dataType: "json",
